Assign "anonymous" values a meaningful id for binding analysis
A lot of values are assigned to named keys with object syntax. For `FunctionExpression`, `ClassExpression` and `ArrowFunctionExpression` types at least, it's super useful to assign property names as identifiers the same way the v8 engine does for stack traces now. This just helps the analyzer to recognize the name bindings that actually exist.